Update test case
The test should test the loveca number calculated is the best or
not, which means an extra loveca doesn’t help. The original test
doesn’t include the estimated number, so it is not testing what we
want to test.

Commit 7b07bdbfaa04bbd41e6fa364ab8c96887c5e2848 includes extra lines.
This comment is similar to it, however, without no extra line